// src/types/rawTransaction.ts
/* LICENSE

© 2023 by KyneSys Labs, licensed under CC BY-NC-ND 4.0

Full license text: https://creativecommons.org/licenses/by-nc-nd/4.0/legalcode
Human readable license: https://creativecommons.org/licenses/by-nc-nd/4.0/

KyneSys Labs: https://www.kynesys.xyz/

*/

export interface RawTransaction {
    blockNumber: number
    signature: string
    status: string
    hash: string
    content: NonNullable<string>
    type: string
    from: any
    to: any
    amount: number
    nonce: number
    timestamp: number
    networkFee: number
    rpcFee: number
    additionalFee: number
}